Update file/folder structure
