+++ src/index.ts
@@ -1,7 +1,9 @@
 import { InteractionType, verifyKey } from 'discord-interactions';
-import { handleRevealStringAction, handleViewLogsAction } from './interactions/button-actions';
-import { handleConfigureCommand, handleStringCommand } from './interactions/commands';
-import { handleStringModalSubmit } from './interactions/modal-submit';
+import { handleRevealStringAction } from './interactions/button-actions/reveal-string';
+import { handleViewLogsAction } from './interactions/button-actions/view-logs';
+import { handleConfigureCommand } from './interactions/commands/configure';
+import { handleStringCommand } from './interactions/commands/string';
+import { handleStringModalSubmit } from './interactions/modal-submit/string-submit';
 import { handlePing } from './interactions/ping';
 
 export default {
@@ -40,7 +42,7 @@ export default {
 			}
 		}
 
-		// 3. Button interactions
+		// 3. Button actions
 		if (interaction.type === InteractionType.MESSAGE_COMPONENT) {
 			const [action, id] = interaction.data.custom_id.split('#');
 			switch (action) {
