+++ src/interactions/commands/configure.ts
@@ -1,33 +1,5 @@
 import { InteractionResponseType } from 'discord-interactions';
-
-export async function handleStringCommand() {
-	return Response.json({
-		type: InteractionResponseType.MODAL,
-		data: {
-			custom_id: 'string_modal',
-			title: 'Create Protected String',
-			components: [
-				{
-					type: 1, // Action Row
-					components: [
-						{
-							type: 4, // Text Input
-							custom_id: 'string_content',
-							label: 'Your String',
-							style: 1,
-							min_length: 1,
-							max_length: 800,
-							placeholder: 'String to protect...',
-							required: true,
-						},
-					],
-				},
-			],
-		},
-	});
-}
-
-export type ChannelSettingsConfig = { ping_enabled?: boolean; ping_role?: string };
+import { ChannelSettingsConfig } from '../types';
 
 export async function handleConfigureCommand({ interaction, db }: { interaction: any; db: D1Database }) {
 	// Get config changes from parameters
