+++ src/interactions/types.ts
@@ -1,0 +1,4 @@
+export interface ChannelSettingsConfig {
+	ping_enabled?: boolean;
+	ping_role?: string;
+}
